+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 `OpenLocate-iOS` adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0](https://github.com/OpenLocate/openlocate-ios/tag/1.1.0)
+
+#### Added
+
+- Add ability to send location data to mulitple URL endpoints
+- Add ability to set the location data transmission interval via `Configuration` class
+- Change default transmission interval to 6 hours instead of 8 hours
+
+#### Fixed
+
+- Fix duplicate location data being sent. (#71)
+
+---
+
 ## [1.0.0](https://github.com/OpenLocate/openlocate-ios/tag/1.0.0)
 
 #### Added
@@ -18,6 +32,6 @@
 
 ---
 
-## [0.1.0]
+## 0.1.0
 
 Initial release!